feat: simpler configuration class
