+++ object-storage-spring-boot-autoconfigure/src/main/java/com/github/anilople/object/storage/config/ObjectStorageAutoConfiguration.java
@@ -18,10 +18,21 @@
 
 package com.github.anilople.object.storage.config;
 
+import java.net.URI;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.AwsCredentials;
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
 /**
  * {@link EnableAutoConfiguration Auto-configuration} for Object Storage support.
@@ -29,6 +40,65 @@ import org.springframework.context.annotation.Import;
  * @author wxq
  */
 @Configuration
-@ConditionalOnProperty(name = "object.storage.enabled", matchIfMissing = true)
-@Import(value = {ObjectStorageConfiguration.class})
-public class ObjectStorageAutoConfiguration {}
+@ConditionalOnClass(S3Client.class)
+@ConditionalOnProperty(name = "object.storage.enabled", havingValue = "true")
+@EnableConfigurationProperties(ObjectStorageProperties.class)
+public class ObjectStorageAutoConfiguration {
+
+  private final ObjectStorageProperties properties;
+
+  public ObjectStorageAutoConfiguration(ObjectStorageProperties properties) {
+    this.properties = properties;
+  }
+
+  public Region resolveRegion() {
+    return Region.of(this.properties.getRegion());
+  }
+
+  public URI resolveEndpoint() {
+    return URI.create(this.properties.getEndpoint());
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public AwsCredentials awsCredentials() {
+    return AwsBasicCredentials.create(
+        this.properties.getAccessKeyId(), this.properties.getSecretAccessKey());
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public AwsCredentialsProvider awsCredentialsProvider(AwsCredentials awsCredentials) {
+    return StaticCredentialsProvider.create(awsCredentials);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public S3Client s3Client(AwsCredentialsProvider awsCredentialsProvider) {
+    final Region region = this.resolveRegion();
+    final URI endpoint = this.resolveEndpoint();
+
+    S3Client s3Client =
+        S3Client.builder()
+            .region(region)
+            .endpointOverride(endpoint)
+            .credentialsProvider(awsCredentialsProvider)
+            .build();
+    return s3Client;
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public S3Presigner s3Presigner(AwsCredentialsProvider awsCredentialsProvider) {
+    final Region region = this.resolveRegion();
+    final URI endpoint = this.resolveEndpoint();
+
+    S3Presigner s3Presigner =
+        S3Presigner.builder()
+            .region(region)
+            .endpointOverride(endpoint)
+            .credentialsProvider(awsCredentialsProvider)
+            .build();
+    return s3Presigner;
+  }
+}
